% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{divmod}
\alias{divmod}
\title{Integer Division and Modulo}
\usage{
divmod(x, div)
}
\arguments{
\item{x}{Numeric vector of values to be divided.}

\item{div}{Numeric vector or scalar divisor. If scalar, it will be
recycled to match the length of \code{x}.}
}
\value{
A list with:
\describe{
\item{quotient}{Integer division results.}
\item{remainder}{Modulo results.}
}
}
\description{
Compute the integer quotient and remainder and return both as a list.
}
\examples{
\donttest{
divmod(17, 5)
divmod(1:10, 3)
divmod(10:12, 1:3)
}

}
