% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{get_last_range}
\alias{get_last_range}
\title{Get the last written table range}
\usage{
get_last_range(wb)
}
\arguments{
\item{wb}{An openxlsx \code{Workbook} object that has been written using
\code{\link[=save_data_wb]{save_data_wb()}}.}
}
\value{
A list with components \code{sheet}, \code{start_cell}, and \code{end_cell}.
}
\description{
Returns the most recently recorded table range stored in
\code{attr(wb, "instead.last_range")}.
}
\seealso{
\code{\link[=save_data_wb]{save_data_wb()}}, \code{\link[=get_last_range_by_sheet]{get_last_range_by_sheet()}}
}
