% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{get_last_range_by_sheet}
\alias{get_last_range_by_sheet}
\title{Get the last written table range for a worksheet}
\usage{
get_last_range_by_sheet(wb, sheet)
}
\arguments{
\item{wb}{An openxlsx \code{Workbook} object that has been written using
\code{\link[=save_data_wb]{save_data_wb()}}.}

\item{sheet}{A single worksheet name.}
}
\value{
A list with components \code{sheet}, \code{start_cell}, and \code{end_cell}.
}
\description{
Returns the most recently recorded table range for a given worksheet stored
in \code{attr(wb, "instead.last_range_by_sheet")}.
}
\seealso{
\code{\link[=save_data_wb]{save_data_wb()}}, \code{\link[=get_last_range]{get_last_range()}}
}
