% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{get_next_cell_left_by_sheet}
\alias{get_next_cell_left_by_sheet}
\title{Get the next cell to the left of the last written table on a worksheet}
\usage{
get_next_cell_left_by_sheet(wb, sheet, col_spacer = 2L)
}
\arguments{
\item{wb}{An openxlsx \code{Workbook} object that has been written using
\code{\link[=save_data_wb]{save_data_wb()}}.}

\item{sheet}{A single worksheet name.}

\item{col_spacer}{Integer (>= 0). Number of blank columns between tables.
Default \code{2L}.}
}
\value{
A numeric vector \code{c(row, col)} representing the next cell position.
}
\description{
Computes the next top-left cell position \strong{to the left} of the most recently
written table on a given worksheet recorded by \code{\link[=save_data_wb]{save_data_wb()}}.
}
\seealso{
\code{\link[=get_last_range_by_sheet]{get_last_range_by_sheet()}}, \code{\link[=get_next_cell_left]{get_next_cell_left()}}
}
