% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_decimal}
\alias{has_decimal}
\title{Tests whether each element of a numeric vector is non-integer
(i.e., has a fractional part) within a given tolerance.}
\usage{
has_decimal(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{A numeric vector to check.}

\item{tol}{Numeric tolerance for floating-point comparison.
Defaults to \code{.Machine$double.eps^0.5}.}
}
\value{
A logical vector of the same length as \code{x}, where \code{TRUE}
indicates that the value has a decimal (fractional) component.
}
\description{
Tests whether each element of a numeric vector is non-integer
(i.e., has a fractional part) within a given tolerance.
}
\details{
The test is based on comparing each value to its nearest integer:
\code{abs(x - round(x)) > tol}.
This accounts for floating-point rounding errors.
}
\examples{
\donttest{
has_decimal(c(1, 2.5, 3.0, 4.75))
#> [1] FALSE  TRUE FALSE  TRUE
}

}
