% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg_fail}
\alias{msg_fail}
\title{Print a failure message}
\usage{
msg_fail(x)
}
\arguments{
\item{x}{Character; message text.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a failure message
}
\examples{
msg_fail("Workbook save failed")

}
