% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg_info}
\alias{msg_info}
\title{Print an info message}
\usage{
msg_info(x)
}
\arguments{
\item{x}{Character; message text.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an info message
}
\examples{
msg_info("Loading workbook")

}
