% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg_rule}
\alias{msg_rule}
\title{Print a section rule}
\usage{
msg_rule(
  x,
  line = 2,
  color = c("cyan", "blue", "red", "green", "yellow", "magenta", "white", "silver")
)
}
\arguments{
\item{x}{Character; section title.}

\item{line}{Integer; rule style passed to \code{cli::rule()}.}

\item{color}{Character; one of \code{"cyan"}, \code{"blue"}, \code{"red"}, \code{"green"},
\code{"yellow"}, \code{"magenta"}, \code{"white"}, or \code{"silver"}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Prints a colored section rule using \code{cli::rule()}.
}
