% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg_step}
\alias{msg_step}
\title{Print a progress step message}
\usage{
msg_step(i, n, x)
}
\arguments{
\item{i}{Current index.}

\item{n}{Total number of steps.}

\item{x}{Character; message text.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Prints a step message such as \verb{[01/10] Plan A}.
}
\examples{
msg_step(1, 10, "Plan A")

}
