% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg_warn}
\alias{msg_warn}
\title{Print a warning message}
\usage{
msg_warn(x)
}
\arguments{
\item{x}{Character; message text.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a warning message
}
\examples{
msg_warn("No rows matched")

}
