% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_class}
\alias{remove_class}
\title{Remove one or more classes from an object}
\usage{
remove_class(x, classes)
}
\arguments{
\item{x}{An object with a class attribute.}

\item{classes}{Character vector of class names to remove.}
}
\value{
The same object \code{x}, with the specified class(es) removed.
}
\description{
Safely removes specified class name(s) from an object's class vector.
Mirrors the internal logic and style of \code{\link[=prepend_class]{prepend_class()}}, but
removes the specified \code{classes} instead of prepending them.
}
\examples{
x <- prepend_class(mtcars, c("custom", "tag"))
class(x)
# [1] "custom" "tag" "data.frame"

y <- remove_class(x, "custom")
class(y)
# [1] "tag" "data.frame"

z <- remove_class(y, c("tag", "data.frame"))
class(z)
# [1] "list"

}
