% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{check_file_overwrite}
\alias{check_file_overwrite}
\title{Check that a file does not already exist}
\usage{
check_file_overwrite(file, overwrite = FALSE)
}
\arguments{
\item{file}{File path.}

\item{overwrite}{Logical; if \code{TRUE}, existing files are allowed.}
}
\value{
Invisibly returns \code{file}.
}
\description{
Stops if \code{file} already exists and \code{overwrite = FALSE}.
}
