% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{check_file_recreate}
\alias{check_file_recreate}
\title{Recreate a file if it already exists}
\usage{
check_file_recreate(file)
}
\arguments{
\item{file}{File path.}
}
\value{
Invisibly returns \code{file}.
}
\description{
If \code{file} exists, asks for confirmation before deleting it.
}
