% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{intersect_cols}
\alias{intersect_cols}
\title{Get Common Column Names Between Two Data Frames}
\usage{
intersect_cols(x, y)
}
\arguments{
\item{x}{A data.frame.}

\item{y}{A data.frame.}
}
\value{
A character vector of column names present in both \code{x} and \code{y}.
}
\description{
Returns the names of columns that exist in both data frames.
Useful for aligning datasets before merging or performing column-wise operations.
}
\examples{
\donttest{
df1 <- data.frame(a = 1:3, b = 4:6, c = 7:9)
df2 <- data.frame(b = 10:12, c = 13:15, d = 16:18)
intersect_cols(df1, df2)
# [1] "b" "c"
}

}
